start refactoring config tab
