+++ node_launcher/gui/menu/manage_bitcoind/bitcoind_configuration_tab.py
@@ -21,22 +21,22 @@ class BitcoindConfigurationTab(QWidget):
 
         self.bitcoin = bitcoin
 
-        self.bitcoin_layout = QVBoxLayout()
+        self.layout = QVBoxLayout()
 
         self.show_bitcoin_conf = QPushButton('Show bitcoin.conf')
         # noinspection PyUnresolvedReferences
         self.show_bitcoin_conf.clicked.connect(
             lambda: reveal(self.node_set.bitcoin.file.directory)
         )
-        self.bitcoin_layout.addWidget(self.show_bitcoin_conf)
+        self.layout.addWidget(self.show_bitcoin_conf)
 
         self.data_directory_group_box = DataDirectoryBox(bitcoin=self.bitcoin)
         self.data_directory_group_box.file_dialog.new_data_directory.connect(
             self.change_datadir
         )
 
-        self.bitcoin_layout.addWidget(self.data_directory_group_box)
-        self.bitcoin_layout.setAlignment(self.data_directory_group_box, Qt.AlignHCenter)
+        self.layout.addWidget(self.data_directory_group_box)
+        self.layout.setAlignment(self.data_directory_group_box, Qt.AlignHCenter)
 
         self.enable_wallet_label = QLabel('Enable wallet')
         self.enable_wallet_widget = QCheckBox('Enable Wallet')
@@ -44,9 +44,9 @@ class BitcoindConfigurationTab(QWidget):
         self.enable_wallet_widget.stateChanged.connect(
             lambda x: self.update_config('disablewallet', not bool(x))
         )
-        self.bitcoin_layout.addWidget(self.enable_wallet_widget)
+        self.layout.addWidget(self.enable_wallet_widget)
 
-        self.setLayout(self.bitcoin_layout)
+        self.setLayout(self.layout)
 
     def change_datadir(self, new_datadir: str):
         self.bitcoin.file['datadir'] = new_datadir
